requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
requests>=2.25.0
python-dotenv>=0.19.0
flask>=2.0.0
joblib>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
openai>=1.0.0